+++ frontend/app/templates/authentication-callback.hbs
@@ -1,0 +1,1 @@
+{{outlet}}
